+++ server/api/board/controller.ts
@@ -5,6 +5,7 @@
 //   get: () => ({ status: 200, body: boardUsecace.getBoard() }),
 //   post: () => ({ status: 201, body: boardUsecace2.startBoard() }),
 // }));
+
 import { boardUseCace } from '$/useCase/boardUseCase';
 import { defineController } from './$relay';
