+++ server/api/index.ts
@@ -7,9 +7,18 @@
 // }>;
 
 import type { DefineMethods } from 'aspida';
+import type { TaskModel } from 'commonTypesWithClient/models';
 
 export type Methods = DefineMethods<{
   get: {
-    resBody: string;
+    query?: {
+      limit?: number;
+    };
+
+    resBody: TaskModel[];
+  };
+  post: {
+    reqBody: Pick<TaskModel, 'label'>;
+    resBody: TaskModel;
   };
 }>;
