+++ server/useCase/aaa.ts
@@ -8,7 +8,7 @@ dotenv.config();
 
 const configuration = new Configuration({
   // apiKey,
-  apiKey: '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',
+  apiKey: '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',
 });
 
 const openai = new OpenAIApi(configuration);
