+++ server/useCase/boardUsecase2.ts
@@ -1,7 +1,6 @@
-// import type { UserId } from '$/commonTypesWithClient/branded';
 // import { func1 } from './aaa';
 
-// let board: BoardArray = [
+// const board: BoardArray = [
 //   [0, 0, 0, 0, 0, 0, 0, 0],
 //   [0, 0, 0, 0, 0, 0, 0, 0],
 //   [0, 0, 0, 0, 3, 0, 0, 0],
@@ -137,7 +136,7 @@
 //   return { board, turn, passCount };
 // };
 
-// const makeBoard = () => {
+// setInterval(() => {
 //   if (onoff === 1) {
 //     console.table(board);
 //     const canselId = setInterval(() => {
@@ -157,31 +156,14 @@
 //           }
 //         });
 //       };
-//       return () => clearInterval(canselId);
 //     });
+//     return () => clearInterval(canselId);
 //   }
-// };
-// makeBoard
+// }, 1000);
+
 // export const boardUsecace2 = {
 //   getBoard: () => board,
 
-//   clickBoard: (params: { x: number; y: number; turn: number }, userId: UserId) => {
-//     return clickBoard({ x: params.x, y: params.y }, board, params.turn, passCount);
-//   },
-//   resetBoard: () => {
-//     board = [
-//       [0, 0, 0, 0, 0, 0, 0, 0],
-//       [0, 0, 0, 0, 0, 0, 0, 0],
-//       [0, 0, 0, 0, 3, 0, 0, 0],
-//       [0, 0, 0, 1, 2, 3, 0, 0],
-//       [0, 0, 3, 2, 1, 0, 0, 0],
-//       [0, 0, 0, 3, 0, 0, 0, 0],
-//       [0, 0, 0, 0, 0, 0, 0, 0],
-//       [0, 0, 0, 0, 0, 0, 0, 0],
-//     ];
-//     onoff = 0;
-//     return board;
-//   },
 //   startBoard: async () => {
 //     onoff = 1;
 //   },
